# dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outrider",
    "description": "Agentic PR review (intake \u2192 triage \u2192 analyze \u21c4 trace \u2192 synthesize \u2192 hitl \u2192 publish).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/slack/install": {
      "get": {
        "tags": [
          "slack-oauth"
        ],
        "summary": "Slack Install",
        "description": "Admin-authed install start: validate channel \u2192 sign state \u2192 hand back the Slack\nauthorize URL. Content-negotiated: `Accept: application/json` (the dashboard\n\"Connect Slack\" flow, which must READ the URL to navigate the browser \u2014 a `fetch`\ncannot read a cross-origin `302`) returns `SlackInstallURL`; anything else keeps the\n`302` redirect (the headless `curl` flow the README documents). Same URL either way.",
        "operationId": "slack_install_slack_install_get",
        "parameters": [
          {
            "name": "installation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Installation Id"
            }
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accept: application/json \u2014 the Slack authorize URL to navigate to.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackInstallURL"
                }
              }
            }
          },
          "302": {
            "description": "Default \u2014 redirect (Location) to Slack's authorize page."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/slack/oauth/callback": {
      "get": {
        "tags": [
          "slack-oauth"
        ],
        "summary": "Slack Oauth Callback",
        "description": "Public OAuth callback: verify state \u2192 exchange code \u2192 encrypt \u2192 persist.\n\nIdentity is read from the VERIFIED state, never from the query params. Failure\nmodes fail closed: denied/bad/expired/forged \u2192 4xx with nothing persisted.",
        "operationId": "slack_oauth_callback_slack_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SlackInstallURL": {
        "properties": {
          "authorize_url": {
            "type": "string",
            "title": "Authorize Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "authorize_url"
        ],
        "title": "SlackInstallURL",
        "description": "JSON body of `GET /slack/install` when the caller sends `Accept: application/json`\n(the dashboard \"Connect Slack\" flow). Carries the Slack authorize URL the browser\nthen navigates to \u2014 the same URL the default `302` puts in `Location`, so the\nsigned `state` is exposed identically. Admin-gated like the redirect."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}